ゲームの初期化を別ファイルに切り出した
